app/backend/knowledge_base_approach/requirements.txt
appdirs==1.4.3
apturl==0.5.2
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
emoji==1.6.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==6.5.1
ipython==7.29.0
isodate==0.6.0
jedi==0.18.1
jupyter-client==7.0.6
jupyter-core==4.9.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
monotonic==1.5
more-itertools==4.2.0
nest-asyncio==1.5.1
netifaces==0.10.4
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
Owlready2==0.35
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.22
protobuf==3.19.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
rdflib==6.0.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
stanford-openie==1.2.0
stanfordnlp==0.2.0
stanza==1.3.0
systemd-python==234
torch==1.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
wget==3.2
xkit==0.0.0
zipp==1.0.0
